+++ src/renderer/RenderPass/Compute/ComputePostProcess/ComputePostProcessPass.h
@@ -13,7 +13,7 @@ namespace vg::renderer
         ComputePostProcessPass();
         ~ComputePostProcessPass();
 
-        void Setup(const gfx::RenderPassContext & _renderContext, double _dt) override;
+        void Setup(const gfx::RenderPassContext & _renderContext, float _dt) override;
         void Render(const gfx::RenderPassContext & _renderContext, gfx::CommandList * _cmdList) const override;
 
     private:
