+++ src/renderer/RenderPass/Render2D/Background/BackgroundPass.h
@@ -13,7 +13,7 @@ namespace vg::renderer
         BackgroundPass();
         ~BackgroundPass();
 
-        void Setup(const gfx::RenderPassContext & _renderContext, double _dt) override;
+        void Setup(const gfx::RenderPassContext & _renderContext, float _dt) override;
         void Render(const gfx::RenderPassContext & _renderContext, gfx::CommandList * _cmdList) const override;
 
     private:
