+++ src/renderer/RenderPass/RenderObjects/DepthOnly/DepthOnlyPass.hpp
@@ -23,7 +23,7 @@ namespace vg::renderer
     //--------------------------------------------------------------------------------------
     // Setup executed each frame, for each pass instance
     //--------------------------------------------------------------------------------------
-    void DepthOnlyPass::Setup(const gfx::RenderPassContext & _renderPassContext, double _dt)
+    void DepthOnlyPass::Setup(const gfx::RenderPassContext & _renderPassContext, float _dt)
     {
         writeRenderTarget(0, _renderPassContext.getFrameGraphID("Color"));
         writeDepthStencil(_renderPassContext.getFrameGraphID("DepthStencil"));
