+++ src/renderer/RenderPass/RenderObjects/Forward/ForwardPass.h
@@ -13,7 +13,7 @@ namespace vg::renderer
         ForwardPass();
         ~ForwardPass();
 
-        void Setup(const gfx::RenderPassContext & _renderContext, double _dt) override;
+        void Setup(const gfx::RenderPassContext & _renderContext, float _dt) override;
         void Render(const gfx::RenderPassContext & _renderContext, gfx::CommandList * _cmdList) const override;
     };
 }
